requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
supabase>=2.9.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
